LLM localRAG/requirements.txt
ollama
faiss-cpu
numpy
sentence-transformers
tensorflow
tf-keras
keras==2.14
transformers
tk
